refactor(plugins): move zellij_exports and friends to a different file
