+++ zellij-server/src/plugins/mod.rs
@@ -1,6 +1,7 @@
 mod plugin_loader;
 mod wasm_bridge;
 mod plugin_map;
+mod zellij_exports;
 use log::info;
 use std::{collections::HashMap, fs, path::PathBuf};
 use wasmer::Store;
